requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-dotenv
requests
langchain
langchain-core
langchain-google-vertexai
langgraph